# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/reports/{id}/resolve": {
      "post": {
        "operationId": "AdminModerationController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "resolution": {
                        "type": "string",
                        "enum": [
                          "DISMISSED"
                        ]
                      },
                      "reason": {
                        "default": "",
                        "type": "string",
                        "maxLength": 500
                      }
                    },
                    "required": [
                      "resolution"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "resolution": {
                        "type": "string",
                        "enum": [
                          "WARNED"
                        ]
                      },
                      "reason": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 500
                      }
                    },
                    "required": [
                      "resolution",
                      "reason"
                    ]
                  },
                  {
                    "type": "object",
                    "properties": {
                      "reason": {
                        "type": "string",
                        "minLength": 3,
                        "maxLength": 500
                      },
                      "days": {
                        "nullable": true,
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 365
                      },
                      "resolution": {
                        "type": "string",
                        "enum": [
                          "SUSPENDED"
                        ]
                      }
                    },
                    "required": [
                      "reason",
                      "days",
                      "resolution"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users": {
      "get": {
        "operationId": "AdminModerationController_users",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "schema": {
              "default": "",
              "type": "string",
              "maxLength": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "pseudo": {
                        "nullable": true,
                        "type": "string"
                      },
                      "email": {
                        "type": "string"
                      },
                      "role": {
                        "type": "string",
                        "enum": [
                          "PLAYER",
                          "VENUE_STAFF",
                          "ADMIN"
                        ]
                      },
                      "minor": {
                        "type": "boolean"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                      },
                      "suspension": {
                        "nullable": true,
                        "type": "object",
                        "properties": {
                          "at": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "until": {
                            "nullable": true,
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "reason": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "at",
                          "until",
                          "reason"
                        ]
                      },
                      "openReports": {
                        "type": "integer",
                        "minimum": -9007199254740991,
                        "maximum": 9007199254740991
                      }
                    },
                    "required": [
                      "id",
                      "pseudo",
                      "email",
                      "role",
                      "minor",
                      "createdAt",
                      "suspension",
                      "openReports"
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{id}": {
      "get": {
        "operationId": "AdminModerationController_user",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "pseudo": {
                      "nullable": true,
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string",
                      "enum": [
                        "PLAYER",
                        "VENUE_STAFF",
                        "ADMIN"
                      ]
                    },
                    "minor": {
                      "type": "boolean"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time",
                      "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                    },
                    "suspension": {
                      "nullable": true,
                      "type": "object",
                      "properties": {
                        "at": {
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "until": {
                          "nullable": true,
                          "type": "string",
                          "format": "date-time",
                          "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                        },
                        "reason": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "at",
                        "until",
                        "reason"
                      ]
                    },
                    "openReports": {
                      "type": "integer",
                      "minimum": -9007199254740991,
                      "maximum": 9007199254740991
                    },
                    "avatarUrl": {
                      "nullable": true,
                      "type": "string"
                    },
                    "avatarStatus": {
                      "nullable": true,
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "APPROVED",
                        "REJECTED"
                      ]
                    },
                    "deleted": {
                      "type": "boolean"
                    },
                    "reports": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string",
                            "enum": [
                              "HARASSMENT",
                              "INAPPROPRIATE_CONTENT",
                              "CHEATING",
                              "NO_SHOW",
                              "MINOR_SAFETY",
                              "OTHER"
                            ]
                          },
                          "details": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "resolvedAt": {
                            "nullable": true,
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "resolution": {
                            "nullable": true,
                            "type": "string",
                            "enum": [
                              "DISMISSED",
                              "WARNED",
                              "SUSPENDED"
                            ]
                          },
                          "reporter": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "pseudo": {
                                "nullable": true,
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "pseudo"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "reason",
                          "details",
                          "createdAt",
                          "resolvedAt",
                          "resolution",
                          "reporter"
                        ]
                      }
                    },
                    "actions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "action": {
                            "type": "string",
                            "enum": [
                              "REPORT_DISMISS",
                              "REPORT_WARN",
                              "REPORT_SUSPEND",
                              "USER_SUSPEND",
                              "USER_UNSUSPEND",
                              "AVATAR_APPROVE",
                              "AVATAR_REJECT",
                              "VENUE_UPDATE",
                              "EVENT_CREATE",
                              "EVENT_UPDATE",
                              "EVENT_CANCEL",
                              "GAME_MERGE"
                            ]
                          },
                          "targetId": {
                            "type": "string"
                          },
                          "reason": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time",
                            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
                          },
                          "admin": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "pseudo": {
                                "nullable": true,
                                "type": "string"
                              }
                            },
                            "required": [
                              "id",
                              "pseudo"
                            ]
                          }
                        },
                        "required": [
                          "id",
                          "action",
                          "targetId",
                          "reason",
                          "createdAt",
                          "admin"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "pseudo",
                    "email",
                    "role",
                    "minor",
                    "createdAt",
                    "suspension",
                    "openReports",
                    "avatarUrl",
                    "avatarStatus",
                    "deleted",
                    "reports",
                    "actions"
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{id}/suspend": {
      "post": {
        "operationId": "AdminModerationController_suspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  },
                  "days": {
                    "nullable": true,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 365
                  }
                },
                "required": [
                  "reason",
                  "days"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/users/{id}/unsuspend": {
      "post": {
        "operationId": "AdminModerationController_unsuspend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
    "/admin/avatars/{id}/approve": {
      "post": {
        "operationId": "AdminModerationController_approveAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
    "/admin/avatars/{id}/reject": {
      "post": {
        "operationId": "AdminModerationController_rejectAvatar",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 500
                  }
                },
                "required": [
                  "reason"
                ]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Kwatro",
    "description": "",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "dev-user": {
        "type": "apiKey",
        "in": "header",
        "name": "x-dev-user-id"
      }
    },
...
  },
  "security": [
    {
      "dev-user": []
    }
  ]
}